Fixed all compiler warnings at current /W levels (Imago)
Fixed MS-SA 973882 (Imago)

Changed some #ifdef'd OutputDebugStrings to ZDebugOutput (Imago)

Updated todo.txt (Imago)

Removed unnecessary FedSrv files (Imago)

Fixed version displayed in server log (per sgt_baker's complaint) (Imago)

Added rel to SlmVer.h for a release number in the build string (Dogbones)

Added extra logging to FedSrv wrt last FAZDev server crash (Imago)

Fixed chat log being created when no registry entry was defined (Imago)

Allegiance client can now use LogToFile & OutputDebugString (Imago)

Prevented loading WMV dlls unnecessarily (Imago)

Fixed all dependencies and default FZRetail flavor for VS2010 (Imago)
Note: AGC/AllSrv/Allegiance are now independent of one another wrt build order.  AGCGuid.lib moved to Guids project!

git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/FAZR5Contrib@486 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/AGC/AGCEventData.cpp
@@ -242,7 +242,7 @@ void CAGCEventData::CopyVariableData(LPCSTR pszContext, LPCOLESTR pszSubject,
       case VT_R4:
       {
         // (...) pushes a float argument onto the stack as a double
-        float var = va_arg(argptr, DOUBLE);
+        float var = (float)va_arg(argptr, DOUBLE);
         COPY_VAR(var);
         break;
       }
